Feature/rework command validate execute test with shm (#1264)
rework command_validate_execute_test with shared_model
